Correct type for tip calculator

+++ Day-2_Data_Types/TipCalculator.py
@@ -12,7 +12,7 @@ print("Welcome to the tip calculator!")
 total_bill = input("What was the total bill? " )
 tip = input("How much tip would you like to give? 10, 12, or 15? ")
 number_of_people = input("How many people to split the bill? ")
-bill_with_tip = float(total_bill) + float(total_bill) * int(total_bill) / 100
+bill_with_tip = float(total_bill) + float(total_bill) * float(total_bill) / 100
 bill_individual = round(bill_with_tip / int(number_of_people), 2)
 
 print(f"Each person should pay: ${bill_individual}")
